+++ bin/free-coding-models.js
@@ -2293,7 +2293,7 @@ async function ensureProxyRunning(fcmConfig, { forceRestart = false } = {}) {
   }
 
   const proxyToken = `fcm_${randomUUID().replace(/-/g, '')}`
-  const proxy = new ProxyServer({ accounts, proxyApiKey: proxyToken })
+  const proxy = new ProxyServer({ accounts, proxyApiKey: proxyToken, upstreamTimeoutMs: 120_000 })
   const { port } = await proxy.start()
   activeProxy = proxy